Error message on Environment page #686

+++ source/src/main/java/org/cerberus/crud/service/impl/CountryEnvDeployTypeService.java
@@ -35,6 +35,7 @@ import org.cerberus.exception.CerberusException;
 import org.cerberus.util.answer.Answer;
 import org.cerberus.util.answer.AnswerItem;
 import org.cerberus.util.answer.AnswerList;
+import org.cerberus.util.answer.AnswerUtil;
 
 @Service
 public class CountryEnvDeployTypeService implements ICountryEnvDeployTypeService {
@@ -97,9 +98,9 @@ public class CountryEnvDeployTypeService implements ICountryEnvDeployTypeService
     @Override
     public Answer compareListAndUpdateInsertDeleteElements(String system, String country, String environement, List<CountryEnvDeployType> newList) {
         Answer ans = new Answer(null);
-        MessageEvent msg = null;
-        msg = new MessageEvent(MessageEventEnum.DATA_OPERATION_OK);
-        msg.setDescription(msg.getDescription().replace("%ITEM%", OBJECT_NAME).replace("%OPERATION%", "UPDATE"));
+
+        MessageEvent msg1 = new MessageEvent(MessageEventEnum.GENERIC_OK);
+        Answer finalAnswer = new Answer(msg1);
 
         List<CountryEnvDeployType> oldList = new ArrayList();
         try {
@@ -120,12 +121,16 @@ public class CountryEnvDeployTypeService implements ICountryEnvDeployTypeService
         for (CountryEnvDeployType objectDifference : listToUpdateOrInsertToIterate) {
             for (CountryEnvDeployType objectInDatabase : oldList) {
                 if (objectDifference.hasSameKey(objectInDatabase)) {
-                    this.update(objectDifference);
+                    ans = this.update(objectDifference);
+                    finalAnswer = AnswerUtil.agregateAnswer(finalAnswer, (Answer) ans);
                     listToUpdateOrInsert.remove(objectDifference);
                 }
             }
         }
-        this.createList(listToUpdateOrInsert);
+        if (!listToUpdateOrInsert.isEmpty()) {
+            ans = this.createList(listToUpdateOrInsert);
+            finalAnswer = AnswerUtil.agregateAnswer(finalAnswer, (Answer) ans);
+        }
 
         /**
          * Iterate on (TestCaseStep From Database - TestCaseStep From Page). If
@@ -143,8 +148,11 @@ public class CountryEnvDeployTypeService implements ICountryEnvDeployTypeService
                 }
             }
         }
-        ans = this.deleteList(listToDelete);
-        return new Answer(msg);
+        if (!listToDelete.isEmpty()) {
+            ans = this.deleteList(listToDelete);
+            finalAnswer = AnswerUtil.agregateAnswer(finalAnswer, (Answer) ans);
+        }
+        return finalAnswer;
     }
 
     @Override
