+++ source/src/main/java/org/cerberus/crud/service/impl/CountryEnvLinkService.java
@@ -33,6 +33,7 @@ import org.cerberus.enums.MessageGeneralEnum;
 import org.cerberus.util.answer.Answer;
 import org.cerberus.util.answer.AnswerItem;
 import org.cerberus.util.answer.AnswerList;
+import org.cerberus.util.answer.AnswerUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -101,9 +102,9 @@ public class CountryEnvLinkService implements ICountryEnvLinkService {
     @Override
     public Answer compareListAndUpdateInsertDeleteElements(String system, String country, String environement, List<CountryEnvLink> newList) {
         Answer ans = new Answer(null);
-        MessageEvent msg = null;
-        msg = new MessageEvent(MessageEventEnum.DATA_OPERATION_OK);
-        msg.setDescription(msg.getDescription().replace("%ITEM%", OBJECT_NAME).replace("%OPERATION%", "UPDATE"));
+
+        MessageEvent msg1 = new MessageEvent(MessageEventEnum.GENERIC_OK);
+        Answer finalAnswer = new Answer(msg1);
 
         List<CountryEnvLink> oldList = new ArrayList();
         try {
@@ -124,12 +125,16 @@ public class CountryEnvLinkService implements ICountryEnvLinkService {
         for (CountryEnvLink objectDifference : listToUpdateOrInsertToIterate) {
             for (CountryEnvLink objectInDatabase : oldList) {
                 if (objectDifference.hasSameKey(objectInDatabase)) {
-                    this.update(objectDifference);
+                    ans = this.update(objectDifference);
+                    finalAnswer = AnswerUtil.agregateAnswer(finalAnswer, (Answer) ans);
                     listToUpdateOrInsert.remove(objectDifference);
                 }
             }
         }
-        this.createList(listToUpdateOrInsert);
+        if (!listToUpdateOrInsert.isEmpty()) {
+            ans = this.createList(listToUpdateOrInsert);
+            finalAnswer = AnswerUtil.agregateAnswer(finalAnswer, (Answer) ans);
+        }
 
         /**
          * Iterate on (TestCaseStep From Database - TestCaseStep From Page). If
@@ -147,8 +152,11 @@ public class CountryEnvLinkService implements ICountryEnvLinkService {
                 }
             }
         }
-        ans = this.deleteList(listToDelete);
-        return new Answer(msg);
+        if (!listToDelete.isEmpty()) {
+            ans = this.deleteList(listToDelete);
+            finalAnswer = AnswerUtil.agregateAnswer(finalAnswer, (Answer) ans);
+        }
+        return finalAnswer;
     }
 
     @Override
