+++ source/src/main/java/org/cerberus/crud/service/impl/CountryEnvironmentDatabaseService.java
@@ -32,6 +32,7 @@ import org.cerberus.enums.MessageGeneralEnum;
 import org.cerberus.util.answer.Answer;
 import org.cerberus.util.answer.AnswerItem;
 import org.cerberus.util.answer.AnswerList;
+import org.cerberus.util.answer.AnswerUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -46,7 +47,7 @@ public class CountryEnvironmentDatabaseService implements ICountryEnvironmentDat
     private ICountryEnvironmentDatabaseDAO countryEnvironmentDatabaseDao;
 
     private final String OBJECT_NAME = "CountryEnvironmentDatabase";
-    
+
     private static final org.apache.log4j.Logger LOG = org.apache.log4j.Logger.getLogger(CountryEnvironmentDatabaseService.class);
 
     @Override
@@ -140,9 +141,9 @@ public class CountryEnvironmentDatabaseService implements ICountryEnvironmentDat
     @Override
     public Answer compareListAndUpdateInsertDeleteElements(String system, String country, String environement, List<CountryEnvironmentDatabase> newList) {
         Answer ans = new Answer(null);
-        MessageEvent msg = null;
-        msg = new MessageEvent(MessageEventEnum.DATA_OPERATION_OK);
-        msg.setDescription(msg.getDescription().replace("%ITEM%", OBJECT_NAME).replace("%OPERATION%", "UPDATE"));
+
+        MessageEvent msg1 = new MessageEvent(MessageEventEnum.GENERIC_OK);
+        Answer finalAnswer = new Answer(msg1);
 
         List<CountryEnvironmentDatabase> oldList = new ArrayList();
         try {
@@ -163,12 +164,16 @@ public class CountryEnvironmentDatabaseService implements ICountryEnvironmentDat
         for (CountryEnvironmentDatabase objectDifference : listToUpdateOrInsertToIterate) {
             for (CountryEnvironmentDatabase objectInDatabase : oldList) {
                 if (objectDifference.hasSameKey(objectInDatabase)) {
-                    this.update(objectDifference);
+                    ans = this.update(objectDifference);
+                    finalAnswer = AnswerUtil.agregateAnswer(finalAnswer, (Answer) ans);
                     listToUpdateOrInsert.remove(objectDifference);
                 }
             }
         }
-        this.createList(listToUpdateOrInsert);
+        if (!listToUpdateOrInsert.isEmpty()) {
+            ans = this.createList(listToUpdateOrInsert);
+            finalAnswer = AnswerUtil.agregateAnswer(finalAnswer, (Answer) ans);
+        }
 
         /**
          * Iterate on (TestCaseStep From Database - TestCaseStep From Page). If
@@ -186,8 +191,11 @@ public class CountryEnvironmentDatabaseService implements ICountryEnvironmentDat
                 }
             }
         }
-        ans = this.deleteList(listToDelete);
-        return new Answer(msg);
+        if (!listToDelete.isEmpty()) {
+            ans = this.deleteList(listToDelete);
+            finalAnswer = AnswerUtil.agregateAnswer(finalAnswer, (Answer) ans);
+        }
+        return finalAnswer;
     }
 
     @Override
